requirements.txt
pandas==2.3.0
numpy==1.24.3
matplotlib==3.10.3
seaborn==0.13.2
scikit-surprise==1.1.4
scikit-learn==1.7.0
torch==2.7.1
dash==3.0.4
